splitted calendar useGetData + serverside:  added adminChecks and caching

+++ server/routes/ethereum/calendars/dropsRoute.js
@@ -1,45 +1,62 @@
-const express = require("express")
-const Drops = require("../../../models/calendars/DropsModel")
+const express = require("express");
+const Drops = require("../../../models/calendars/DropsModel");
+const checkAdmin = require("../../../middleware/checkAdmin");
+const checkAuth = require("../../../middleware/checkAuth");
 
-const dropsRoute = express()
+const dropsRoute = express();
+
+const { lruCache } = require("../../../services/cache/lru");
 
 dropsRoute.get("/", async (req, res) => {
-
+    const ttl = 6 * 60 * 60 * 1000;
+    
     try {
+        const drops = await Drops.find({});
+        if (!drops) throw new Error("Drops not found");
 
-        const drops = await Drops.find({});;
+        await lruCache(Drops.find({}), {
+            key: `drops`,
+            ttl,
+        });
 
-        if (!drops) throw new Error("Drops not found")
-
-        res.status(200).json({drops})
+        res.status(200).json({drops});
     }
     catch (e) {
-        res.status(400).json({ error: e, status: "error" })
+        res.status(400).json({ error: e, status: "error" });
     }
 
 })
 
-dropsRoute.post('/', async (req, res) => {
-
-    const { collectionName, links, price, supply, timestamp, address, eventDescription } = req.body || {};
+dropsRoute.post('/', checkAuth, checkAdmin, async (req, res) => {
+    const { address } = req.userDetails;
+    const isAdmin = req.isAdmin;
+    const { collectionName, links, price, supply, timestamp, eventDescription } = req.body || {};
 
     try {
-        const drops = await Drops.create({ collectionName, links, price, supply, timestamp, address, eventDescription })
-        if (!drops) throw Error('Something went wrong saving the admin drops event');
-        return res.status(200).json({drops});
+        if (isAdmin) {
+            const drops = await Drops.create({ collectionName, links, price, supply, timestamp, eventDescription })
+            if (!drops) throw Error('Something went wrong saving the admin drops event');
+            return res.status(200).json({drops});
+        } else {
+            throw Error('You are not authorized');
+        }
     } catch (e) {
         return res.status(400).json({ msg: e.message });
     }
 });
 
-dropsRoute.delete('/', async (req, res) => {
-
+dropsRoute.delete('/', checkAuth, checkAdmin, async (req, res) => {
+    const isAdmin = req.isAdmin;
     const { id } = req.body || {};
 
     try {
-        const drops = await Drops.deleteOne({_id: id});
-        if (!drops) throw Error('Something went wrong deleting all the admin drops event');
-        return res.status(200).json({drops});
+        if (isAdmin) {
+            const drops = await Drops.deleteOne({_id: id});
+            if (!drops) throw Error('Something went wrong deleting all the admin drops event');
+            return res.status(200).json({drops});
+        } else {
+            throw Error('You are not authorized');
+        }
     } catch (e) {
         return res.status(400).json({ msg: e.message });
     }
